제작한 가짜 WantedService와 실제 WantedService합침
